+++ packages/plugins/auto-doc/src/ui/components/views/template-overview/template-overview.svelte
@@ -114,7 +114,9 @@ function openFileSelectorIfNotMasterTemplate() {
 }
 
 let templatesConvertedToTableRow = $derived(
-	allTemplates.map(mapElementToTableRow)
+	allTemplates
+		.map(mapElementToTableRow)
+		.sort((a, b) => b.lastEdited.getTime() - a.lastEdited.getTime())
 )
 $effect(() => {
 	docTemplatesStore.setMasterTemplateFlag(isMasterTemplate)